requirements.txt
Flask==3.0.3
openai-whisper==20240927
